revert: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@59728 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/glib20/libglib/gio/gapplicationcommandline.c
@@ -666,8 +666,8 @@ const gchar *
 g_application_command_line_getenv (GApplicationCommandLine *cmdline,
                                    const gchar             *name)
 {
-  size_t length = strlen (name);
-  size_t i;
+  gint length = strlen (name);
+  gint i;
 
   /* TODO: expand on windows */
   if (cmdline->priv->environ)
