+++ src/router/glib20/libglib/gio/gdummyfile.c
@@ -239,7 +239,7 @@ static const char *
 match_prefix (const char *path, 
               const char *prefix)
 {
-  size_t prefix_len;
+  int prefix_len;
 
   prefix_len = strlen (prefix);
   if (strncmp (path, prefix, prefix_len) != 0)
