+++ src/router/glib20/libglib/girepository/gifieldinfo.c
@@ -190,7 +190,7 @@ gi_field_info_get_field (GIFieldInfo *field_info,
                          void        *mem,
                          GIArgument  *value)
 {
-  size_t offset;
+  int offset;
   GITypeInfo *type_info;
   gboolean result = FALSE;
 
@@ -385,7 +385,7 @@ gi_field_info_set_field (GIFieldInfo      *field_info,
                          void             *mem,
                          const GIArgument *value)
 {
-  size_t offset;
+  int offset;
   GITypeInfo *type_info;
   gboolean result = FALSE;
